Moved project content and tasks components under portal/project folder
